fix: make getStickyAdjacentToCenterPane work
- remove bug that occur when user was on sticky cells
